refactor: move selection & upgrade methods into host classes
